fix: optimizacion de layouts
